Store extension installation details in report
This report allows us to see what the environment was when the AppSignal
extension was installed, so we can compare it to the diagnose report
generated by `mix appsignal.diagnose` afterwards.

Basic information, such as information about Elixir, the chosen agent
and extension build.

Combine the information from the `install.log` and
`appsignal.architecture` files in one file that stores more easily
parseable metadata.

It's especially good to know if the `LDFLAGS` were set or not during
installation. Helps track down if it was correctly installed for this
issue:
https://docs.appsignal.com/support/known-issues/compilation-dependencies-required-at-runtime.html